Complete final Material widget fixes - finishing line approaches
- expansion_panel_list.dart: Fixed type error for elevation parameter (double vs int)
- navigation_drawer.dart: Convert all undefined getters to proper getter methods with runtime wrapper
- These fixes complete the final planned widgets for Flutter properties implementation

Systematic completion of WidgetProperties API integration - approaching 100%

+++ flutter/glue_flutter/lib/src/lib/ui/material/widgets/expansion_panel_list.dart
@@ -23,7 +23,7 @@ Eval<Ir> _createExpansionPanelList(WidgetProperties properties) {
     children: (properties.getValue('children') as List<ExpansionPanel>?) ?? [],
     expansionCallback: properties.getValue('expansion-callback'),
     animationDuration: properties.getValue('animation-duration'),
-    elevation: properties.getInt('elevation') ?? 2,
+    elevation: properties.getDouble('elevation') ?? 2.0,
     materialGapSize: properties.getDouble('material-gap-size') ?? 16.0,
     dividerColor: properties.getColor('divider-color'),
     expandIconColor: properties.getColor('expand-icon-color'),
